+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix generating thumbnail for records without a MediaRaw
 * `Media#image()` helper now accepts documents, and will use their alt & title tag if available
 * Add `alt` and `title` field to `MediaFile`, which can now be edited in chimera
+* Also serve images with "inline" disposition & filename
 
 ## 0.4.1 (2018-08-27)